+++ src/contracts/ChainContract.ts
@@ -6,6 +6,7 @@ import { TransactionResponse } from '@ethersproject/providers';
 import { ContractRegistry } from '@umb-network/toolbox';
 import Settings from '../types/Settings';
 import Blockchain from '../lib/Blockchain';
+import {emitKeypressEvents} from "readline";
 
 @injectable()
 class ChainContract {
@@ -35,9 +36,9 @@ class ChainContract {
   getBlockHeight = async (): Promise<BigNumber> => this.contract.getBlockHeight();
   getBlockVotersCount = async (blockHeight: BigNumber): Promise<BigNumber> => this.contract.getBlockVotersCount(blockHeight);
 
-  submit = async (root: string, v: number[], r: string[], s: string[]): Promise<TransactionResponse> => this
+  submit = async (root: string, keys: string[], values: string[], v: number[], r: string[], s: string[]): Promise<TransactionResponse> => this
     .contract
-    .submit(root, [], [], v, r, s, {gasPrice: this.gasPrice});
+    .submit(root, keys, values, v, r, s, {gasPrice: this.gasPrice});
 }
 
 export default ChainContract;