+++ src/services/FeedSynchronizer.ts
@@ -1,17 +1,15 @@
 import { getModelForClass } from '@typegoose/typegoose';
-import axios from 'axios';
 import { injectable } from 'inversify';
 import { v4 as uuid } from 'uuid';
-import { JSONPath } from 'jsonpath-plus';
 import Leaf from '../models/Leaf';
 import Feed from '../models/Feed';
+import FeedValueResolver from "./FeedValueResolver";
 
 @injectable()
 class FeedSynchronizer {
   async apply(feed: Feed): Promise<Leaf[]> {
-    const response = await axios.get(feed.sourceUrl);
     const leaf = this.buildLeaf(feed)
-    const value = this.extractValue(response.data, feed.valuePath);
+    const value = await FeedValueResolver.apply(feed);
 
     if (value) {
       leaf.value = value;
@@ -29,12 +27,6 @@ class FeedSynchronizer {
     leaf.feedId = feed._id;
     leaf.label = feed.leafLabel;
     return leaf;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  private extractValue = (data: any, valuePath: string): number | undefined => {
-    const match: number[] | null = JSONPath({ json: data, path: valuePath });
-    if (match) return match[0];
   }
 
   private save = async (leaf: Leaf): Promise<void> => {